backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
PyYAML>=6.0.1
